requirements.txt
gunicorn
django-heroku
django-simple-captcha
mock
flake8
black
coverage
coveralls
Faker